* Update: Updated code to use the queue index retrieved from SAI * Update: Allow queue-deletion * Update: Miscellaneous performance improvements and optimizations * Update: Enable modification of buffer profile after creation * Update: Disallow removal scheduler group if it is being used * Update: Revise the threshold range value * Update: SAI 1.0 scheduler group and queue changes * Update: SAI 1.0 QoS-queue related change * Update: Miscellaneous performance improvements and optimizations * Feature: Add support for 100g/50g/25g interface support * Bugfix: Miscellaneous bug fixes * Cleanup: Miscellaneous cleanup

+++ scripts/lib/python/nas_qos_buffer_profile.py
@@ -27,7 +27,7 @@ class BufferProfileCPSObj:
         'pool-id': ('leaf', 'uint64_t'),
         'buffer-size': ('leaf', 'uint32_t'),
         'threshold-mode': ('leaf', 'enum', 'base-qos:buffer-threshold-mode:'),
-        'shared-dynamic-threshold': ('leaf', 'uint32_t'),
+        'shared-dynamic-threshold': ('leaf', 'uint8_t'),
         'shared-static-threshold': ('leaf', 'uint32_t'),
         'xoff-threshold': ('leaf', 'uint32_t'),
         'xon-threshold': ('leaf', 'uint32_t'),
